GretaAnalogWriter/requirements.txt
azure-iot-device~=2.0.0
pymodbus
numpy
aenum
pyparsing